# Vectorize output for reduction