@zasdfgbnm 2020-04-24
@ngimel This is ready except for ROCm failures. I will open a separate PR to split the compilation unit to fix ROCm.